+++ src/main/java/lotto/domain/LottoResultStatistics.java
@@ -1,9 +1,9 @@
 package lotto.domain;
 
-import lotto.domain.winnerlotto.rank.LottoRank;
-import lotto.domain.lottoticket.LottoTickets;
+import lotto.domain.lottos.rank.LottoRank;
+import lotto.domain.lottos.LottoTickets;
 import lotto.domain.money.Money;
-import lotto.domain.winnerlotto.LottoWinner;
+import lotto.domain.lottos.winnerlotto.LottoWinner;
 
 import java.util.Arrays;
 import java.util.EnumMap;
