+++ src/test/java/lotto/domain/LottoWinnerBonusNumberTest.java
@@ -1,8 +1,8 @@
 package lotto.domain;
 
-import lotto.domain.winnerlotto.LottoWinnerTicket;
-import lotto.domain.number.LottoNumber;
-import lotto.domain.winnerlotto.LottoWinnerBonusNumber;
+import lotto.domain.lottos.winnerlotto.LottoWinnerTicket;
+import lotto.domain.lottos.LottoNumber;
+import lotto.domain.lottos.winnerlotto.LottoWinnerBonusNumber;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
