+++ src/test/java/lotto/domain/LottoWinnerTest.java
@@ -1,9 +1,9 @@
 package lotto.domain;
 
-import lotto.domain.winnerlotto.LottoWinnerTicket;
-import lotto.domain.number.LottoNumber;
-import lotto.domain.winnerlotto.LottoWinnerBonusNumber;
-import lotto.domain.winnerlotto.LottoWinner;
+import lotto.domain.lottos.winnerlotto.LottoWinnerTicket;
+import lotto.domain.lottos.LottoNumber;
+import lotto.domain.lottos.winnerlotto.LottoWinnerBonusNumber;
+import lotto.domain.lottos.winnerlotto.LottoWinner;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
